Struct method changes, more tests, README update
Adding methods to structs is now separated into creating a method
with a StructSpec as a receiver, and adding a MethodSpec to a struct
to have it appear next to the struct when the StructSpec is stringified.
This should make it more obvious what attaching a method is for, as opposed
to creating a standalone method with the struct as a receiver.